feature: Create List users service/moved loggers
