+++ apps/server/src/modules/users/services/ListUsersService.ts
@@ -1,0 +1,14 @@
+import { getCustomRepository } from 'typeorm'
+import User from '../typeorm/entities/User'
+import UsersRepository from '../typeorm/repositories/UsersRepository'
+
+class ListUsersService {
+  public async execute(): Promise<User[]> {
+    const usersRepo = getCustomRepository(UsersRepository)
+    const users = await usersRepo.find()
+
+    return users
+  }
+}
+
+export default ListUsersService
