+++ apps/server/src/modules/users/typeorm/entities/User.ts
@@ -17,7 +17,7 @@ class User {
   @Column()
   email: string
 
-  @Column()
+  @Column({ select: false })
   password: string
 
   @Column('boolean')
